Fix some okapi-ir tests
